+++ tt_metal/api/tt-metalium/experimental/tensor/README.md
@@ -18,6 +18,22 @@ This folder
 Headers in this folder are not in the experimental folder as this staging area is meant to be short lived,
 most of the concepts hosted are well tested, production code already.
 
+## Header Mapping
+
+The following headers were migrated from TTNN to this directory. Forward headers remain in the original TTNN locations for backwards compatibility.
+
+| TTNN Source | TT-Metal Destination |
+|-------------|----------------------|
+| `ttnn/api/ttnn/tensor/types.hpp` | `tensor_types.hpp` |
+| `ttnn/api/ttnn/tensor/tensor_spec.hpp` | `spec/tensor_spec.hpp` |
+| `ttnn/api/ttnn/tensor/layout/alignment.hpp` | `spec/layout/alignment.hpp` |
+| `ttnn/api/ttnn/tensor/layout/layout.hpp` | `spec/layout/layout.hpp` |
+| `ttnn/api/ttnn/tensor/layout/page_config.hpp` | `spec/layout/page_config.hpp` |
+| `ttnn/api/ttnn/tensor/layout/tensor_layout.hpp` | `spec/layout/tensor_layout.hpp` |
+| `ttnn/api/ttnn/tensor/memory_config/memory_config.hpp` | `spec/memory_config/memory_config.hpp` |
+| `ttnn/api/ttnn/distributed/distributed_configs.hpp` | `topology/distributed_tensor_configs.hpp` |
+| `ttnn/api/ttnn/distributed/tensor_topology.hpp` | `topology/tensor_topology.hpp` |
+
 ## Life-time
 
 This folder is expected to be short-lived, the effort is tracked by:
