+++ tt_metal/api/tt-metalium/experimental/tensor/tensor_types.hpp
@@ -1,0 +1,101 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <cstdint>
+#include <memory>
+#include <optional>
+#include <variant>
+#include <vector>
+#include <algorithm>
+
+#include <tt-metalium/bfloat16.hpp>
+#include <tt-metalium/core_coord.hpp>
+#include <tt-metalium/buffer.hpp>
+#include <tt-metalium/mesh_buffer.hpp>
+#include <tt-metalium/device.hpp>
+#include <tt-metalium/tt_backend_api_types.hpp>
+#include <tt_stl/span.hpp>
+#include <tt-metalium/shape.hpp>
+
+namespace tt::tt_metal {
+
+static constexpr std::uint8_t VERSION_ID = 5;
+
+enum class DataType {
+    BFLOAT16 = 0,
+    FLOAT32 = 1,
+    UINT32 = 2,
+    BFLOAT8_B = 3,
+    BFLOAT4_B = 4,
+    UINT8 = 5,
+    UINT16 = 6,
+    INT32 = 7,
+    INVALID = 8,
+};
+
+std::ostream& operator<<(std::ostream& os, const tt::tt_metal::DataType& data_type);
+
+template <typename T>
+consteval DataType convert_to_data_type() {
+    if constexpr (std::is_same_v<T, uint8_t>) {
+        return DataType::UINT8;
+    } else if constexpr (std::is_same_v<T, uint16_t>) {
+        return DataType::UINT16;
+    } else if constexpr (std::is_same_v<T, int32_t>) {
+        return DataType::INT32;
+    } else if constexpr (std::is_same_v<T, uint32_t>) {
+        return DataType::UINT32;
+    } else if constexpr (std::is_same_v<T, float>) {
+        return DataType::FLOAT32;
+    } else if constexpr (std::is_same_v<T, ::bfloat16>) {
+        return DataType::BFLOAT16;
+    } else {
+        static_assert(tt::stl::concepts::always_false_v<T>, "Unsupported DataType!");
+    }
+}
+
+bool is_floating_point(DataType dtype);
+
+bool is_block_float(DataType dtype);
+
+// Specifies Tensor storage type.
+enum class StorageType {
+    HOST = 0,
+    DEVICE = 1,
+};
+
+tt::DataFormat datatype_to_dataformat_converter(DataType datatype);
+tt::tt_metal::DataType dataformat_to_datatype_converter(tt::DataFormat dataformat);
+
+static constexpr std::size_t MAX_NUM_DIMENSIONS = 8;
+
+using Array1D = std::array<uint32_t, 1>;
+using Array2D = std::array<uint32_t, 2>;
+using Array3D = std::array<uint32_t, 3>;
+using Array4D = std::array<uint32_t, 4>;
+using Array5D = std::array<uint32_t, 5>;
+using Array6D = std::array<uint32_t, 6>;
+using Array7D = std::array<uint32_t, 7>;
+using Array8D = std::array<uint32_t, 8>;
+
+struct NdShardSpec {
+    Shape shard_shape;
+    CoreRangeSet grid;
+    ShardOrientation orientation = ShardOrientation::ROW_MAJOR;
+    ShardDistributionStrategy shard_distribution_strategy = ShardDistributionStrategy::ROUND_ROBIN_1D;
+
+    NdShardSpec with_shard_shape(Shape new_shard_shape) const {
+        return NdShardSpec{std::move(new_shard_shape), grid, orientation, shard_distribution_strategy};
+    }
+
+    bool operator==(const NdShardSpec& other) const = default;
+    bool operator!=(const NdShardSpec& other) const = default;
+};
+
+using PadValue = std::variant<uint32_t, float>;
+std::ostream& operator<<(std::ostream& os, const NdShardSpec& spec);
+
+}  // namespace tt::tt_metal
