+++ ttnn/api/ttnn/tensor/layout/layout.hpp
@@ -4,12 +4,4 @@
 
 #pragma once
 
-#include <ostream>
-
-namespace tt::tt_metal {
-
-enum class Layout { ROW_MAJOR = 0, TILE = 1, INVALID = 2 };
-
-std::ostream& operator<<(std::ostream& os, const tt::tt_metal::Layout& layout);
-
-}  // namespace tt::tt_metal
+#include <tt-metalium/experimental/tensor/spec/layout/layout.hpp>
